contract address shows in contact list and not send flow
